Commit post Chapter 10 and 11 on Django for Beginners
Project updates after working with Bootstrap and User Account Log In, Sign Up, Password Change and Password Reset.  A slight error is shown that I'm unable to fix when request for a "Password Reset" if user is unable to login.  Comment is shown on line 42 in "shop/templates/layout.html".