Finished the book Django for Begineers
The last commit to the branch after reading through the text "Django for Beginners."  This was a tremendous book.  I have learned lots from it.  Currently facing a problem with the application when creating new comments of an item or referencing existing comments with one item.  We will check to see what errors I'm running into and talk through the coding aspect in our upcoming session.